dashboard_streamlit/requirements.txt
streamlit
requests
pandas
psycopg2-binary
plotly
python-dotenv
requests>=2.28.0
